move controller request and response to model folder
